+++ src/consensus/pbft/pbftconfig.h
@@ -3,7 +3,7 @@
 #pragma once
 #include "consensus/pbft/libbyz/libbyz.h"
 #include "consensus/pbft/libbyz/pbft_assert.h"
-#include "consensus/pbft/pbftmetadata.h"
+#include "consensus/pbft/pbftrequests.h"
 #include "enclave/rpcmap.h"
 #include "pbftdeps.h"
